Refactor discriminator
