+++ src/main/java/com/networknt/schema/ValidationContext.java
@@ -17,15 +17,11 @@
 package com.networknt.schema;
 
 import java.net.URI;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Stack;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.networknt.schema.SpecVersion.VersionFlag;
 import com.networknt.schema.uri.URIFactory;
 import com.networknt.schema.urn.URNFactory;
@@ -36,7 +32,6 @@ public class ValidationContext {
     private final JsonMetaSchema metaSchema;
     private final JsonSchemaFactory jsonSchemaFactory;
     private SchemaValidatorsConfig config;
-    private final Stack<DiscriminatorContext> discriminatorContexts = new Stack<>();
     private final ConcurrentMap<String, JsonSchema> schemaReferences = new ConcurrentHashMap<>();
     private final ConcurrentMap<String, JsonSchema> schemaResources = new ConcurrentHashMap<>();
 
@@ -116,21 +111,6 @@ public class ValidationContext {
         return this.schemaResources;
     }
 
-    public DiscriminatorContext getCurrentDiscriminatorContext() {
-        if (!this.discriminatorContexts.empty()) {
-            return this.discriminatorContexts.peek();
-        }
-        return null; // this is the case when we get on a schema that has a discriminator, but it's not used in anyOf
-    }
-
-    public void enterDiscriminatorContext(final DiscriminatorContext ctx, @SuppressWarnings("unused") JsonNodePath instanceLocation) {
-        this.discriminatorContexts.push(ctx);
-    }
-
-    public void leaveDiscriminatorContextImmediately(@SuppressWarnings("unused") JsonNodePath instanceLocation) {
-        this.discriminatorContexts.pop();
-    }
-
     public JsonMetaSchema getMetaSchema() {
         return this.metaSchema;
     }
@@ -138,40 +118,5 @@ public class ValidationContext {
     public Optional<VersionFlag> activeDialect() {
         String metaSchema = getMetaSchema().getUri();
         return SpecVersionDetector.detectOptionalVersion(metaSchema);
-    }
-
-    public static class DiscriminatorContext {
-        private final Map<String, ObjectNode> discriminators = new HashMap<>();
-
-        private boolean discriminatorMatchFound = false;
-
-        public void registerDiscriminator(final SchemaLocation schemaLocation, final ObjectNode discriminator) {
-            this.discriminators.put("#" + schemaLocation.getFragment().toString(), discriminator);
-        }
-
-        public ObjectNode getDiscriminatorForPath(final SchemaLocation schemaLocation) {
-            return this.discriminators.get("#" + schemaLocation.getFragment().toString());
-        }
-
-        public ObjectNode getDiscriminatorForPath(final String schemaLocation) {
-            return this.discriminators.get(schemaLocation);
-        }
-
-        public void markMatch() {
-            this.discriminatorMatchFound = true;
-        }
-
-        public boolean isDiscriminatorMatchFound() {
-            return this.discriminatorMatchFound;
-        }
-
-        /**
-         * Returns true if we have a discriminator active. In this case no valid match in anyOf should lead to validation failure
-         *
-         * @return true in case there are discriminator candidates
-         */
-        public boolean isActive() {
-            return !this.discriminators.isEmpty();
-        }
     }
 }
